frontend: move conversion amount to its own component
Moving to components so that ConversionAmount can be re-used in
other components.
