+++ frontends/web/src/components/amount/conversion-amount.module.css
@@ -1,0 +1,23 @@
+.txConversionAmount {
+  font-variant: tabular-nums;
+}
+
+.txPrefix,
+.txUnit {
+  color: var(--color-secondary);
+  font-size: 14px;
+  line-height: 1.285714;
+}
+
+.txConversionAmount .txUnit {
+  font-size: 12px;
+}
+
+.txSmallInlineIcon img {
+  max-height: 15px;
+  max-width: 15px;
+  vertical-align: text-bottom;
+}
+.txConversionAmount .txSmallInlineIcon {
+  padding-right: var(--space-eight);
+}
